requirements.txt
# Generated requirements.txt from pyproject.toml dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
pydantic>=2.10.6
httpx>=0.28.1
azure-identity>=1.15.0
azure-cosmos>=4.5.1
httpx-sse>=0.4.0
semantic-kernel>=1.30.0
a2a-sdk[sqlite]>=0.2.9
sqlalchemy>=2.0.41
aiosqlite>=0.21.0
jwcrypto>=1.5.6
pyjwt>=2.10.1
sse-starlette>=2.2.1
starlette>=0.46.1
typing-extensions>=4.12.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1

# Development dependencies
pytest>=8.3.5
pytest-asyncio>=0.21.1
pytest-mock>=3.14.0
ruff>=0.11.2
